refactor: 이벤트 host 라우트를 eventId에서 eventCode로 변경
dashboard/moderation 페이지가 필요로 하는 API가 eventId가 아니라
eventCode를 쓰는데, 라우트 폴더는 [eventId]였다. 이 불일치를 BE와
확인한 결과, 라우트 쪽을 eventCode로 맞추기로 결정했다.

- app/(host)/events/[eventId]/ → app/(host)/events/[eventCode]/로 변경
- dashboard, moderation 페이지는 URL의 eventCode를 그대로 API에 사용
  (GET /events/{eventCode}/feedbacks, GET /admin/feedbacks?eventCode=...&toxic=true)
- moderation의 eventCode 필터는 BE가 2026-08-05에 GET /admin/feedbacks에
  새로 추가한 쿼리 파라미터다. 이전에는 이벤트별로 거를 방법이 API에
  없었는데, 이번에 해결됐다.
- 이벤트 상세 페이지(page.tsx)는 수정·삭제·세션·리포트생성 API가 여전히
  eventId를 요구하므로, GET /events/{eventCode}로 먼저 id를 조회해야
  한다는 점을 주석으로 남겼다.

Refs #5
