+++ app/(host)/events/[eventCode]/dashboard/page.tsx
@@ -1,0 +1,7 @@
+const DashboardPage = () => {
+  // API: GET /events/{eventCode}/feedbacks. URL의 eventCode를 그대로 사용합니다.
+  // /live 화면과 같은 useLiveFeedback 훅을 재사용할 예정입니다.
+  return <div>실시간 모니터링 대시보드 (준비 중)</div>;
+};
+
+export default DashboardPage;
