--- app/(host)/events/[eventId]/moderation/page.tsx
@@ -1,6 +1,0 @@
-const ModerationPage = () => {
-  // API: GET /admin/feedbacks?toxic=true, PATCH /admin/feedbacks/{feedbackId}/hide.
-  return <div>모더레이션 큐 (준비 중)</div>;
-};
-
-export default ModerationPage;
